A lot of things, check description
1. Formatting
2. Add Node.JS Versions to choose from v14 up to v16
3. Help command improved
4. Confirmations
5. Small patches
!!!Was not tested and will soon!!!

+++ src/Commands/DeleteMySQLUser.php
@@ -7,17 +7,23 @@ use Illuminate\Console\Command;
 class DeleteMySQLUser extends Command
 {
 
-  protected $signature = 'phpmyadmin:user:delete';
-  protected $description = 'Create an user account for phpMyAdmin';
+    protected $signature = 'phpmyadmin:user:delete';
+    protected $description = 'Create an user account for phpMyAdmin';
 
-  public function handle()
-  {
-    $this->delete();
-  }
+    public function handle()
+    {
+        $this->delete();
+    }
 
-  private function delete()
-  {
-    exec('mysql -u root -e DROP USER \'phpmyadmin\'@\'%\';');
-    return $this->info('User has been deleted');
-  }
+    private function delete()
+    {
+        if (!$this->confirm('Are you sure you want to delete the user?')) {
+            $this->warn('The user was not deleted');
+
+            return;
+        }
+
+        exec('mysql -u root -e DROP USER \'phpmyadmin\'@\'%\';');
+        return $this->info('User has been deleted');
+    }
 }
